requirements.txt
# CSIRO Mentor - Python Dependencies
# For Azure App Service deployment

fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
gunicorn==21.2.0
